requirements.txt
Django==6.0.2
psycopg2-binary==2.9.11
openai==2.17.0
python-dotenv
dj-database-url
requests
beautifulsoup4
feedparser>=6.0.0
markdown